+++ src/Desktop.Core/Adapters/Etp12Extender.cs
@@ -162,16 +162,14 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// </summary>
         public void CloseSession()
         {
-            if (Session?.CanHandle<ICoreClient>() ?? false)
+            if ((Session?.CanHandle<ICoreClient>()).GetValueOrDefault())
             {
-                Session?.Handler<ICoreClient>()
-                    ?.CloseSession();
+                Session?.Handler<ICoreClient>()?.CloseSession();
             }
 
-            if (Session?.CanHandle<ICoreServer>() ?? false)
+            if ((Session?.CanHandle<ICoreServer>()).GetValueOrDefault())
             {
-                Session?.Handler<ICoreServer>()
-                    ?.CloseSession();
+                Session?.Handler<ICoreServer>()?.CloseSession();
             }
         }
 
@@ -232,6 +230,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="minMessageInterval"></param>
         public void Start(int maxDataItems, int minMessageInterval)
         {
+            if (!Session.IsRegistered<IChannelStreamingConsumer>()) return;
+
             Session.Handler<IChannelStreamingConsumer>()
                 .Start(maxDataItems, minMessageInterval);
         }
@@ -242,6 +242,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uris">The URIs.</param>
         public void ChannelDescribe(IList<string> uris)
         {
+            if (!Session.IsRegistered<IChannelStreamingConsumer>()) return;
+
             Session.Handler<IChannelStreamingConsumer>()
                 .ChannelDescribe(uris);
         }
@@ -253,6 +255,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="startIndex">The start index.</param>
         public void ChannelStreamingStart(IList<ChannelMetadataViewModel> channels, object startIndex)
         {
+            if (!Session.IsRegistered<IChannelStreamingConsumer>()) return;
+
             // Prepare ChannelStreamingInfos startIndexes
             _channelStreamingInfos.Clear();
 
@@ -271,6 +275,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="channelIds"></param>
         public void ChannelStreamingStop(IList<long> channelIds)
         {
+            if (!Session.IsRegistered<IChannelStreamingConsumer>()) return;
+
             Session.Handler<IChannelStreamingConsumer>()
                 .ChannelStreamingStop(channelIds);
         }
@@ -283,6 +289,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="endIndex">The end index.</param>
         public void ChannelRangeRequest(IList<long> channelIds, long startIndex, long endIndex)
         {
+            if (!Session.IsRegistered<IChannelStreamingConsumer>()) return;
+
             var rangeInfo = new ChannelRangeInfo
             {
                 ChannelId = channelIds,
@@ -306,6 +314,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="dataChanges">if set to <c>true</c> supports data changes.</param>
         public void OpenChannelResponse(IMessageHeader request, string uri, long id, Guid uuid, object lastIndex = null, bool infill = true, bool dataChanges = true)
         {
+            if (!Session.IsRegistered<IChannelDataLoadConsumer>()) return;
+
             Session.Handler<IChannelDataLoadConsumer>()
                 .OpenChannelResponse(request, uri, id, uuid, lastIndex, dataChanges);
         }
@@ -317,6 +327,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <returns>The message identifier.</returns>
         public long GetResources(string uri)
         {
+            if (!Session.IsRegistered<IDiscoveryCustomer>()) return 0;
+
             return Session.Handler<IDiscoveryCustomer>()
                 .GetResources(uri);
         }
@@ -328,6 +340,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <returns>The message identifier.</returns>
         public long FindResources(string uri)
         {
+            if (!Session.IsRegistered<IDiscoveryQueryCustomer>()) return 0;
+
             return Session.Handler<IDiscoveryQueryCustomer>()
                 .FindResources(uri);
         }
@@ -338,6 +352,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uri">The URI.</param>
         public void FindObjects(string uri)
         {
+            if (!Session.IsRegistered<IStoreQueryCustomer>()) return;
+
             Session.Handler<IStoreQueryCustomer>()
                 .FindObjects(uri);
         }
@@ -348,6 +364,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uri">The URI.</param>
         public void GetObject(string uri)
         {
+            if (!Session.IsRegistered<IStoreCustomer>()) return;
+
             Session.Handler<IStoreCustomer>()
                 .GetObject(uri);
         }
@@ -358,6 +376,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uri">The URI.</param>
         public void DeleteObject(string uri)
         {
+            if (!Session.IsRegistered<IStoreCustomer>()) return;
+
             Session.Handler<IStoreCustomer>()
                 .DeleteObject(uri);
         }
@@ -374,6 +394,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="childCount">The child count.</param>
         public void PutObject(string uri, string uuid, string name, string xml, string contentType, ResourceTypes resourceType = ResourceTypes.DataObject, int childCount = -1)
         {
+            if (!Session.IsRegistered<IStoreCustomer>()) return;
+
             var dataObject = new DataObject
             {
                 Resource = new Resource
@@ -404,6 +426,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="objectTypes">The object types.</param>
         public void NotificationRequest(string uri, string uuid, long startTime, bool includeObjectData, IList<string> objectTypes)
         {
+            if (!Session.IsRegistered<IStoreNotificationCustomer>()) return;
+
             var request = new NotificationRequestRecord
             {
                 Uri = uri,
@@ -425,6 +449,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uuid"></param>
         public void CancelNotification(string uuid)
         {
+            if (!Session.IsRegistered<IStoreNotificationCustomer>()) return;
+
             var request = _notificationRequests.FirstOrDefault(x => x.Uuid.EqualsIgnoreCase(uuid));
             if (request == null) return;
 
@@ -440,6 +466,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uri">The URI.</param>
         public void FindParts(string uri)
         {
+            if (!Session.IsRegistered<IGrowingObjectQueryCustomer>()) return;
+
             Session.Handler<IGrowingObjectQueryCustomer>()
                 .FindParts(uri);
         }
@@ -451,6 +479,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uid">The UID.</param>
         public void GetPart(string uri, string uid)
         {
+            if (!Session.IsRegistered<IGrowingObjectCustomer>()) return;
+
             Session.Handler<IGrowingObjectCustomer>()
                 .GetPart(uri, uid);
         }
@@ -465,6 +495,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="depthDatum">The depth datum.</param>
         public void GetPartsByRange(string uri, double? startIndex, double? endIndex, string uom, string depthDatum)
         {
+            if (!Session.IsRegistered<IGrowingObjectCustomer>()) return;
+
             Session.Handler<IGrowingObjectCustomer>()
                 .GetPartsByRange(uri, startIndex, endIndex, uom, depthDatum);
         }
@@ -479,6 +511,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="compress"><c>true</c> if the data object should be compressed; otherwise, <c>false</c>.</param>
         public void PutPart(string uri, string uid, string contentType, string data, bool compress)
         {
+            if (!Session.IsRegistered<IGrowingObjectCustomer>()) return;
+
             var dataObject = new DataObject();
             dataObject.SetString(data, compress);
 
@@ -493,6 +527,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="uid">The UID.</param>
         public void DeletePart(string uri, string uid)
         {
+            if (!Session.IsRegistered<IGrowingObjectCustomer>()) return;
+
             Session.Handler<IGrowingObjectCustomer>()
                 .DeletePart(uri, uid);
         }
@@ -507,6 +543,8 @@ namespace PDS.WITSMLstudio.Desktop.Core.Adapters
         /// <param name="depthDatum">The depth datum.</param>
         public void DeletePartsByRange(string uri, double? startIndex, double? endIndex, string uom, string depthDatum)
         {
+            if (!Session.IsRegistered<IGrowingObjectCustomer>()) return;
+
             Session.Handler<IGrowingObjectCustomer>()
                 .DeletePartsByRange(uri, startIndex, endIndex, uom, depthDatum);
         }
